+++ tests/unit/services/test_internal.py
@@ -2,7 +2,7 @@ from unittest import mock
 
 from localstack.constants import VERSION
 from localstack.http import Request
-from localstack.services.internal import CloudFormationUi, HealthResource
+from localstack.services.internal import HealthResource
 from localstack.services.plugins import ServiceManager, ServiceState
 
 
@@ -65,12 +65,3 @@ class TestHealthResource:
             },
             "version": VERSION,
         }
-
-
-class TestCloudFormationUiResource:
-    def test_get(self):
-        resource = CloudFormationUi()
-        response = resource.on_get(Request("GET", "/", body=b"None"))
-        assert response.status == "200 OK"
-        assert "</html>" in response.get_data(as_text=True), "deploy UI did not render HTML"
-        assert "text/html" in response.headers.get("content-type", "")
